policy for bookmarks

+++ app/controllers/bookmarks_controller.rb
@@ -3,6 +3,7 @@ class BookmarksController < ApplicationController
 
   def index
     @user = User.find(params[:user_id])
+    authorize @user, policy_class: BookmarkPolicy
     @bookmarks = @user.bookmarks
   end
 
@@ -10,6 +11,7 @@ class BookmarksController < ApplicationController
     user = User.find(params[:user_id])
     job = Job.find(params[:job_id])
     bookmark = user.bookmarks.new(job_id: params[:job_id])
+    authorize bookmark, policy_class: BookmarkPolicy
     if bookmark.valid? && bookmark.save
       flash[:success] = 'Job Saved'
     else
@@ -22,6 +24,7 @@ class BookmarksController < ApplicationController
   def destroy
     user = User.find(params[:user_id])
     bookmark = Bookmark.find(params[:id])
+    authorize bookmark, policy_class: BookmarkPolicy
     bookmark.destroy
     flash[:success] = 'Saved Job Deleted.'
     redirect_to user_bookmarks_path(user)
